check balance before creating giveaway

+++ handlers/giveaways.js
@@ -2,7 +2,7 @@ const fs = require('fs');
 const config = require("../config.json");
 
 module.exports = {
-  executeCommand: async function (giveawaysData, client, message, command, args) {
+  executeCommand: async function (giveawaysData, walletsData, client, message, command, args) {
     if (args[0] === "help") {
 
       fs.readFile('./templates/help_giveaways.msg', 'utf8', function (err, source) {
@@ -17,33 +17,41 @@ module.exports = {
       let amount = parseFloat(args[3]);
       let description = args.slice(4).join(" ");
 
-      const giveawayEmbed = {
-        color: 0x0099ff,
-        title: description,
-        url: 'https://discord.js.org',
-        author: {
-          name: 'Conceal Network',
-          icon_url: 'https://conceal.network/images/branding/logo.png',
-          url: 'https://discord.gg/YbpHVSd'
-        },
-        description: `React with \:tada: to enter. Prize is ${amount} CCX`,
-        timestamp: new Date(),
-        footer: {
-          text: `${winners} winners | ends at ...`,
-          icon_url: 'https://conceal.network/images/branding/logo.png'
+      walletsData.getBalance(message.member.user.id).then(balanceData => {
+        if (balanceData.balance > (amount * config.metrics.coinUnits)) {
+          const giveawayEmbed = {
+            color: 0x0099ff,
+            title: description,
+            url: 'https://discord.js.org',
+            author: {
+              name: 'Conceal Network',
+              icon_url: 'https://conceal.network/images/branding/logo.png',
+              url: 'https://discord.gg/YbpHVSd'
+            },
+            description: `React with \:tada: to enter. Prize is ${amount} CCX`,
+            timestamp: new Date(),
+            footer: {
+              text: `${winners} winners | ends at ...`,
+              icon_url: 'https://conceal.network/images/branding/logo.png'
+            }
+          };
+
+          // delete the bot command 
+          message.delete().catch(O_o => { });
+
+          message.channel.send({ embed: giveawayEmbed }).then(newMsg => {
+            giveawaysData.createGiveaway(message.member.user.id, newMsg.id, timespan, winners, amount, description).then(data => {
+              newMsg.react('🎉');
+            }).catch(err => {
+              newMsg.delete().catch(O_o => { });
+              message.channel.send(`Error creating giveaway: ${err}`);
+            });
+          });
+        } else {
+          message.channel.send(`insuficient balance ${balanceData.balance / config.metrics.coinUnits} CCX`);
         }
-      };
-
-      // delete the bot command 
-      message.delete().catch(O_o => { });
-
-      message.channel.send({ embed: giveawayEmbed }).then(newMsg => {
-        giveawaysData.createGiveaway(message.member.user.id, newMsg.id, timespan, winners, amount, description).then(data => {
-          newMsg.react('🎉');
-        }).catch(err => {
-          newMsg.delete().catch(O_o => { });
-          message.channel.send(`Error creating giveaway: ${err}`);
-        });
+      }).catch(err => {
+        message.channel.send(err);
       });
     }
   }
